+++ include/seqan3/alignment/matrix/detail/alignment_score_matrix_one_column.hpp
@@ -16,7 +16,7 @@
 #include <seqan3/alignment/matrix/detail/alignment_score_matrix_one_column_base.hpp>
 #include <seqan3/alignment/matrix/detail/alignment_score_matrix_proxy.hpp>
 #include <seqan3/core/concept/core_language.hpp>
-#include <seqan3/core/simd/concept.hpp>
+#include <seqan3/utility/simd/concept.hpp>
 #include <seqan3/std/iterator>
 #include <seqan3/std/ranges>
 #include <seqan3/std/span>
