+++ include/seqan3/alignment/matrix/detail/alignment_score_matrix_one_column_base.hpp
@@ -17,7 +17,7 @@
 #include <vector>
 
 #include <seqan3/core/concept/core_language.hpp>
-#include <seqan3/core/simd/concept.hpp>
+#include <seqan3/utility/simd/concept.hpp>
 #include <seqan3/range/container/aligned_allocator.hpp>
 
 namespace seqan3::detail
