+++ include/seqan3/alignment/pairwise/detail/policy_scoring_scheme.hpp
@@ -13,7 +13,7 @@
 #pragma once
 
 #include <seqan3/alignment/configuration/align_config_scoring_scheme.hpp>
-#include <seqan3/core/simd/concept.hpp>
+#include <seqan3/utility/simd/concept.hpp>
 
 namespace seqan3::detail
 {
