+++ include/seqan3/alignment/pairwise/policy/simd_affine_gap_policy.hpp
@@ -21,7 +21,7 @@
 #include <seqan3/alignment/pairwise/detail/alignment_algorithm_state.hpp>
 #include <seqan3/core/concept/core_language.hpp>
 #include <seqan3/core/configuration/configuration.hpp>
-#include <seqan3/core/simd/concept.hpp>
+#include <seqan3/utility/simd/concept.hpp>
 #include <seqan3/core/simd/simd_algorithm.hpp>
 #include <seqan3/core/simd/simd_traits.hpp>
