+++ include/seqan3/alignment/scoring/detail/simd_match_mismatch_scoring_scheme.hpp
@@ -17,7 +17,7 @@
 #include <seqan3/alphabet/concept.hpp>
 #include <seqan3/core/bit_manipulation.hpp>
 #include <seqan3/core/concept/cereal.hpp>
-#include <seqan3/core/simd/concept.hpp>
+#include <seqan3/utility/simd/concept.hpp>
 #include <seqan3/core/simd/simd_algorithm.hpp>
 #include <seqan3/std/concepts>
