+++ include/seqan3/core/simd/all.hpp
@@ -31,7 +31,7 @@
  * \todo Describe more what SIMD is and how to use it.
  */
 
-#include <seqan3/core/simd/concept.hpp>
+#include <seqan3/utility/simd/concept.hpp>
 #include <seqan3/core/simd/debug_stream_simd.hpp>
 #include <seqan3/core/simd/simd_algorithm.hpp>
 #include <seqan3/core/simd/simd_traits.hpp>
