+++ include/seqan3/core/simd/detail/simd_algorithm_avx512.hpp
@@ -14,7 +14,7 @@
 
 #include <array>
 
-#include <seqan3/core/simd/concept.hpp>
+#include <seqan3/utility/simd/concept.hpp>
 #include <seqan3/core/simd/detail/builtin_simd_intrinsics.hpp>
 #include <seqan3/core/simd/detail/builtin_simd.hpp>
 #include <seqan3/core/simd/simd_traits.hpp>
