+++ include/seqan3/core/simd/simd_algorithm.hpp
@@ -15,7 +15,7 @@
 #include <array>
 #include <utility>
 
-#include <seqan3/core/simd/concept.hpp>
+#include <seqan3/utility/simd/concept.hpp>
 #include <seqan3/core/simd/detail/builtin_simd.hpp>
 #include <seqan3/core/simd/detail/simd_algorithm_sse4.hpp>
 #include <seqan3/core/simd/detail/simd_algorithm_avx2.hpp>
