+++ include/seqan3/core/simd/view_iota_simd.hpp
@@ -15,7 +15,7 @@
 #include <seqan3/std/ranges>
 
 #include <seqan3/core/concept/core_language.hpp>
-#include <seqan3/core/simd/concept.hpp>
+#include <seqan3/utility/simd/concept.hpp>
 #include <seqan3/core/simd/simd_algorithm.hpp>
 #include <seqan3/core/simd/simd_traits.hpp>
