+++ test/include/seqan3/test/simd_utility.hpp
@@ -14,7 +14,7 @@
 
 #include <array>
 
-#include <seqan3/core/simd/concept.hpp>
+#include <seqan3/utility/simd/concept.hpp>
 
 //!\cond DEV
 /*!\brief #SIMD_EQ checks if the sizes and the content of two given
