+++ test/unit/alignment/matrix/alignment_optimum_test.cpp
@@ -11,7 +11,7 @@
 
 #include <seqan3/alignment/matrix/alignment_optimum.hpp>
 #include <seqan3/core/simd/simd.hpp>
-#include <seqan3/core/simd/concept.hpp>
+#include <seqan3/utility/simd/concept.hpp>
 
 #include <seqan3/test/simd_utility.hpp>
