+++ test/unit/alignment/matrix/detail/two_dimensional_matrix_test.cpp
@@ -14,7 +14,7 @@
 #include <seqan3/alignment/matrix/detail/two_dimensional_matrix_iterator_concept.hpp>
 #include <seqan3/alignment/matrix/detail/two_dimensional_matrix.hpp>
 #include <seqan3/alignment/matrix/matrix_concept.hpp>
-#include <seqan3/core/simd/concept.hpp>
+#include <seqan3/utility/simd/concept.hpp>
 #include <seqan3/core/simd/simd.hpp>
 #include <seqan3/core/simd/simd_algorithm.hpp>
 #include <seqan3/test/simd_utility.hpp>
