+++ test/unit/core/simd/detail/builtin_simd_test.cpp
@@ -10,7 +10,7 @@
 #include <seqan3/std/type_traits>
 
 #include <seqan3/core/detail/pack_algorithm.hpp>
-#include <seqan3/core/simd/concept.hpp>
+#include <seqan3/utility/simd/concept.hpp>
 #include <seqan3/core/simd/detail/builtin_simd.hpp>
 #include <seqan3/utility/type_list/type_list.hpp>
